% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-wateryear.r
\name{assert_wy_spec}
\alias{assert_wy_spec}
\title{Assert Water Year Specification}
\usage{
assert_wy_spec(start, end)
}
\arguments{
\item{start, end}{The water year start and end. Can be a \code{Date}
object or a character string.}
}
\description{
Assert the validity of the water year specification.
}
\keyword{internal}
