% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateryear.r
\name{fixed_yday}
\alias{fixed_yday}
\title{Fixed Day of the Year}
\usage{
fixed_yday(x)
}
\arguments{
\item{x}{a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate,
xts, its, ti, jul, timeSeries, or fts object.}
}
\value{
A vector of days of year (numeric vector).
}
\description{
Return the day of the year in the range [1, 366]. Day 60 is skipped
for non-leap years.
}
\keyword{internal}
