% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-year.r
\name{referenceyear}
\alias{referenceyear}
\title{Reference Water Year}
\usage{
referenceyear()
}
\value{
The reference water year (numeric).
}
\description{
Get the reference water year.
}
\keyword{internal}
