% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-year.r
\name{set_referenceyear}
\alias{set_referenceyear}
\title{Set Reference Water Year}
\usage{
set_referenceyear(reference_year, verbose = TRUE)
}
\arguments{
\item{reference_year}{The reference water year.}

\item{verbose}{If \code{TRUE}, produce a message echoing the reference
water year setting.}
}
\description{
Set the reference water year for the current session.
}
\seealso{
\code{\link[=set_wateryear]{set_wateryear()}} \code{\link[=with_referenceyear]{with_referenceyear()}}
}
