% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-wateryear.r
\name{set_wateryear}
\alias{set_wateryear}
\title{Set Water Year}
\usage{
set_wateryear(start = "10-01", end = "09-30", fmt = "\%m-\%d", verbose = TRUE)
}
\arguments{
\item{start, end}{The water year start and end. Can be a \code{Date}
object or a character string.}

\item{fmt}{The date format of the character string. See
\code{\link[=strptime]{strptime()}} for more information on format specifications.}

\item{verbose}{If \code{TRUE}, produce a message echoing the water year
setting.}
}
\description{
Set the water year locale for the current session.
}
