% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateryear.r
\name{wateryear}
\alias{wateryear}
\title{Water Year}
\usage{
wateryear(x)
}
\arguments{
\item{x}{A vector of dates or datetimes.}
}
\value{
A numeric vector of water years.
}
\description{
Get the water year of a calendar date.
}
