% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-year.r
\name{with_referenceyear}
\alias{with_referenceyear}
\title{With Reference Year}
\usage{
with_referenceyear(x)
}
\arguments{
\item{x}{A vector of dates or datetimes.}
}
\value{
The vector \code{x}, with the year overridden with the reference
water year.
}
\description{
Adjust dates or datetimes to the reference year. This is useful for
plotting and summarizing data by day of year.
}
