% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-wateryear.r
\name{set_wateryear}
\alias{set_wateryear}
\title{Set Water Year}
\usage{
set_wateryear(
  start = "0000-10-01",
  end = "0001-09-30",
  fmt = "\%Y-\%m-\%d",
  verbose = TRUE
)
}
\arguments{
\item{start, end}{The water year start and end. Can be a \code{Date}
object or a character string.}

\item{fmt}{The date format of the character string. See
\code{\link[=strptime]{strptime()}} for more information on format specifications.}

\item{verbose}{If \code{TRUE}, produce a message echoing the water year
setting.}
}
\description{
Set the water year locale for the current session.
}
\details{
To set the water year automatically when attaching the
package, specify the option \code{wateryear.default}:

\code{options(wateryear.default = list(start = "10-01", end = "09-30", fmt = "\%m-\%d"))}
}
